Logout fix
